+++ server/requirements.txt
@@ -1,24 +1,27 @@
-h5py~=3.7.0
-numpy~=1.23.4
-SQLAlchemy~=1.4.43
-fastapi~=0.86.0
-pydantic~=1.10.2
-rasterio~=1.3.3
-starlette~=0.20.4
-sqlalchemy-json~=0.5.0
-Rtree~=1.0.1
-geojson-pydantic~=0.4.3
-inquirerpy~=0.3.4
-jsonschema~=4.17.0
-tabulate~=0.9.0
-pandapower~=2.12.1
-pandas~=2.0.1
-pyproj~=3.5.0
-setuptools~=65.5.1
-geopandas~=0.13.0
-shapely~=2.0.1
-ezdxf~=1.1.0
-geojson~=3.0.1
+h5py
+numpy
+SQLAlchemy
+fastapi
+pydantic
+rasterio
+starlette
+sqlalchemy-json
+Rtree
+geojson-pydantic
+inquirerpy
+jsonschema
+tabulate
+pandapower
+pandas
+pyproj
+setuptools
+geopandas
+shapely
+ezdxf
+geojson
 numpy-stl
 saas @ git+https://github.com/cooling-singapore/saas-middleware@8654f3892e49c854dbe056560bfb02226e05a059
-sentry-sdk==1.4.3
+sentry-sdk
+pytest
+python-dotenv
+
