--- server/tests/data/duplicate_geometries/rotated
@@ -1,9 +1,0 @@
-{
-"type": "FeatureCollection",
-"name": "rotated",
-"features": [
-{ "type": "Feature", "properties": { "id": 1122234, "name": "test_building_2", "height": 20.0, "area": 2310.0, "building_type": "commercial:2", "zone_id": 42 }, "geometry": { "type": "Polygon", "coordinates": [ [ [ 103.844813126561533, 1.273863878023676 ], [ 103.844560533308808, 1.273612026226246 ], [ 103.844308677128765, 1.273864619485408 ], [ 103.844561271247841, 1.274116475482259 ], [ 103.844813126561533, 1.273863878023676 ] ] ] } },
-{ "type": "Feature", "id": 7093, "properties": { "id": 10002, "name": "Unknown-97572", "height": 10.0, "area": 2382.8556349952537, "building_type": "commercial:2", "zone_id": 41 }, "geometry": { "type": "Polygon", "coordinates": [ [ [ 103.8437233, 1.2740144 ], [ 103.843721, 1.2740005 ], [ 103.843732, 1.2739987 ], [ 103.8437345, 1.2740071 ], [ 103.8437401, 1.2740199 ], [ 103.8437485, 1.2740286 ], [ 103.8437575, 1.2740355 ], [ 103.8437675, 1.2740389 ], [ 103.8437897, 1.2740416 ], [ 103.8437892, 1.274045 ], [ 103.8439847, 1.2740609 ], [ 103.8440233, 1.2736769 ], [ 103.8438193, 1.2736705 ], [ 103.8438195, 1.2736739 ], [ 103.8438108, 1.2736714 ], [ 103.8438019, 1.2736702 ], [ 103.8437869, 1.273672 ], [ 103.8437719, 1.273678 ], [ 103.8437621, 1.2736856 ], [ 103.8437554, 1.2736968 ], [ 103.8437524, 1.2737109 ], [ 103.8437535, 1.2737218 ], [ 103.8437526, 1.2737294 ], [ 103.843748, 1.2737367 ], [ 103.843747, 1.2737305 ], [ 103.8437223, 1.2737346 ], [ 103.8437206, 1.2737243 ], [ 103.8437047, 1.2737269 ], [ 103.843707, 1.2737411 ], [ 103.8436831, 1.273745 ], [ 103.843682, 1.2737386 ], [ 103.8436333, 1.2737466 ], [ 103.8436346, 1.2737546 ], [ 103.8436249, 1.2737591 ], [ 103.8436193, 1.2737648 ], [ 103.8436157, 1.2737743 ], [ 103.8436136, 1.2737618 ], [ 103.8435087, 1.273779 ], [ 103.8435108, 1.2737915 ], [ 103.8434931, 1.2737944 ], [ 103.8434877, 1.2737615 ], [ 103.8434519, 1.2737674 ], [ 103.8434563, 1.2737941 ], [ 103.8434085, 1.2738019 ], [ 103.8434114, 1.2738195 ], [ 103.8433924, 1.2738226 ], [ 103.8433929, 1.2738258 ], [ 103.8433902, 1.2738252 ], [ 103.843377, 1.2738279 ], [ 103.843367, 1.2738343 ], [ 103.8433601, 1.2738453 ], [ 103.8433568, 1.2738605 ], [ 103.8433567, 1.273868 ], [ 103.8433589, 1.273878 ], [ 103.8433459, 1.2738759 ], [ 103.843339, 1.2739525 ], [ 103.8433877, 1.2740123 ], [ 103.8434816, 1.27402 ], [ 103.8434817, 1.2740176 ], [ 103.8434931, 1.2740207 ], [ 103.8435076, 1.2740219 ], [ 103.8435247, 1.2740206 ], [ 103.8435311, 1.2740217 ], [ 103.8435949, 1.274012 ], [ 103.8435981, 1.2740154 ], [ 103.8436025, 1.2740172 ], [ 103.8436081, 1.2740176 ], [ 103.8436126, 1.2740163 ], [ 103.8436168, 1.274013 ], [ 103.8436244, 1.2740084 ], [ 103.8436312, 1.2740065 ], [ 103.8436385, 1.274009 ], [ 103.8436466, 1.2740152 ], [ 103.843655, 1.2740177 ], [ 103.8436633, 1.2740177 ], [ 103.8436711, 1.2740146 ], [ 103.8436741, 1.2740122 ], [ 103.8436757, 1.2740222 ], [ 103.8437233, 1.2740144 ] ] ] } },
-{ "type": "Feature", "properties": { "id": 1122233, "name": "test_building_1", "height": 20.0, "area": 2310.0, "building_type": "commercial:2", "zone_id": 42 }, "geometry": { "type": "Polygon", "coordinates": [ [ [ 103.84438255, 1.2736859 ], [ 103.84438255, 1.274042596901017 ], [ 103.84473925, 1.2740426 ], [ 103.844739252362103, 1.27368589952042 ], [ 103.84438255, 1.2736859 ] ] ] } }
-]
-}
